Migrate to Svelte 5 & refactor class page (#56)
Migrate to Svelte 5 as well as redesign and refactor the giant mess that was the class page + hypothetical grade calculation code. It's now split into two files and a bunch of functions. Some miscellaneous fixes as well.

* Initial Svelte 5 migration
The class grade page hasn't been migrated to runes mode yet, and the slot syntax is still being used because snippets are broken with flowbite components.

* Giant class page refactor

* Fix vulnerable regex

* Improve grade calculation

* Enormous grade calc / types refactor

* Packages update

* Remove Report Cards page
It's been broken for a while now. You can just use Documents instead. I've left the type definitions just in case.

* Change on: handlers to svelte 5 syntax