Fix regression in NumberInput two-way binding (#120)
When NumberInput was updated to prevent value from possibly being null, the two-way binding created to facilitate it only took the value inputted initially. If it was ever changed externally, it did not reflect it because its internal value was not bound. This fixes it with a janky usage of $effect. Hopefully I can find a less-janky way to do the two-way binding necessary in this situation in the future.