Redesign seen assignments feature (#160)
This shows seen assignments in a more legible left-highlighted border with a new badge indicator as well. It moves the "mark as seen" button to a floating position at the bottom of the screen to make it harder to miss. When logging in on a new device it will no longer mark all existing assignments as new.